feat(layout): move desktop nav to component directory
