+++ libs/ui-kit/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/ui-kit';
 export * from './lib/components';
+export * from './lib/models';
