+++ libs/ui-kit/src/lib/components/desktop-nav/desktop-nav.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import DesktopNav from './desktop-nav';
+
+describe('DesktopNav', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<DesktopNav />);
+    expect(baseElement).toBeTruthy();
+  });
+});
