+++ libs/ui-kit/src/lib/models/index.ts
@@ -1,0 +1,1 @@
+export * from './ui-kit.models';
