+++ libs/ui-kit/src/lib/models/ui-kit.models.ts
@@ -1,0 +1,7 @@
+export interface NavigationItem {
+  label: string;
+  href: string;
+  value?: string;
+}
+
+export type NavigationItems = NavigationItem[];
